+++ app/controllers/admin/app_types_controller.rb
@@ -50,16 +50,21 @@ class Admin::AppTypesController < AdminController
 
   def show
     app_type = Admin::AppType.find(params[:id])
-    f = params[:export_format]
-    if f.present?
-      raise FphsException, "Invalid export format. Allowed formats: #{ValidFormats}" unless f.in? ValidFormats
+
+    exp_format = params[:export_format]
+    show_components = params[:show_components].present?
+    if exp_format.present?
+      raise FphsException, "Invalid export format. Allowed formats: #{ValidFormats}" unless exp_format.in? ValidFormats
+    elsif show_components
+      @app_type = app_type
+      render 'admin/app_types/components'
+      return
     else
-      f = 'json'
+      exp_format = 'json'
     end
 
     app_type.current_admin = current_admin
-
-    send_data app_type.export_config(format: f.to_sym), filename: "#{app_type.name}_config.#{f}"
+    send_data app_type.export_config(format: exp_format.to_sym), filename: "#{app_type.name}_config.#{exp_format}"
   end
 
   def export_migrations