+++ app/models/option_configs/report_options.rb
@@ -21,6 +21,7 @@ module OptionConfigs
 
     configure :list_options, with: %i[hide_in_list list_description]
     configure :view_css, with: %i[classes selectors]
+    configure :criteria_css, with: %i[classes selectors]
     configure :component, with: [:options]
     configure :column_options, with: %i[tags classes hide show_as alt_column_header]